Refactored resample
